auth-service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic[email]
python-dotenv
PyJWT
python-jose[cryptography]